Move average movement and movement direction classes to dedicated sub packages
