+++ src/test/java/org/mastodon/mamut/feature/branch/movement/average/BranchAverageMovementFeatureComputerTest.java
@@ -26,12 +26,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package org.mastodon.mamut.feature.branch.movement;
+package org.mastodon.mamut.feature.branch.movement.average;
 
 import org.junit.Test;
 import org.mastodon.feature.FeatureProjection;
 import org.mastodon.mamut.feature.FeatureComputerTestUtils;
 import org.mastodon.mamut.feature.branch.exampleGraph.ExampleGraph2;
+import org.mastodon.mamut.feature.branch.movement.average.BranchAverageMovementFeature;
 import org.mastodon.mamut.model.branch.BranchSpot;
 import org.scijava.Context;
