+++ src/test/java/org/mastodon/mamut/feature/branch/movement/direction/BranchMovementDirectionFeatureTest.java
@@ -1,4 +1,4 @@
-package org.mastodon.mamut.feature.branch.movement;
+package org.mastodon.mamut.feature.branch.movement.direction;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -8,6 +8,7 @@ import org.mastodon.mamut.feature.AbstractFeatureTest;
 import org.mastodon.mamut.feature.FeatureComputerTestUtils;
 import org.mastodon.mamut.feature.FeatureSerializerTestUtils;
 import org.mastodon.mamut.feature.branch.exampleGraph.ExampleGraph1;
+import org.mastodon.mamut.feature.branch.movement.direction.BranchMovementDirectionFeature;
 import org.mastodon.mamut.model.branch.BranchSpot;
 import org.scijava.Context;
